rename package in generator
