+++ bin/generator/proccesing/folder_manager.go
@@ -1,4 +1,4 @@
-package generator
+package proccesing
 
 import (
 	"fmt"
