+++ bin/generator/proccesing/generation_thub_template.go
@@ -1,4 +1,4 @@
-package generator
+package proccesing
 
 import (
 	"io/ioutil"
