+++ bin/generator/proccesing/generator.go
@@ -1,4 +1,4 @@
-package generator
+package proccesing
 
 import (
 	"encoding/json"
