+++ bin/generator/proccesing/parsing_resource_file.go
@@ -1,4 +1,4 @@
-package generator
+package proccesing
 
 import (
 	"io/ioutil"
