+++ bin/generator/proccesing/parsing_tf_files.go
@@ -1,4 +1,4 @@
-package generator
+package proccesing
 
 import (
 	"encoding/json"
